requirements.txt
streamlit
google-generativeai
localtunnel
